requirements.txt
python-dotenv==1.1.0
llama-index==0.12.37
llama-index-llms-groq==0.3.1
ollama==0.4.8
 llama-index-embeddings-huggingface==0.5.4
streamlit==1.45.1
